Refactors for more natural api
All encryption and distribution is done with file paths (passing a path
to the encrypt function and getting a path back. Distributing a file path.
All fetching and decrypting is done with file objects (fetching a url
returns a file object which can be passed straight into decryption to get
the verified plaintext